- Confirmation Dialog for Delete and Logout Operations. - Notification Dialog for Add and Edit user operations.

+++ src/app/user/user.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import { HomeComponent } from '../home/home.component';
 import { User } from './user';
 import { UserModalComponent } from './user-modal/user-modal.component';
 import { UserService } from './user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user',
@@ -32,12 +32,39 @@ export class UserComponent implements OnInit {
 
     modalDialog.componentInstance.updatedUsers.subscribe((emittedValue) => {
       this.userService.publishValue(emittedValue);
+
+      Swal.fire({
+        position: 'top-end',
+        icon: 'success',
+        title: 'User Edited',
+        showConfirmButton: false,
+        timer: 1500
+      });
     });
   }
 
   deleteUser(user: User): void {
-    this.userService.deleteUser(user).subscribe((users) => {
-      this.userService.publishValue(users);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You won\'t be able to revert this!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Yes, delete it!'
+    })
+    .then((result) => {
+      if (result.isConfirmed) {
+        this.userService.deleteUser(user).subscribe((users) => {
+          this.userService.publishValue(users);
+
+          Swal.fire(
+            'Deleted!',
+            'User has been deleted.',
+            'success'
+          );
+        });
+      }
     });
   }
 }
